ljj新增了一个转账界面，并修改了之前的几个界面

+++ src/main/java/ui/DepositMoney.java
@@ -51,11 +51,10 @@ public class DepositMoney extends JDialog {
     /**返回新增的存款数目
      *
      */
-        Integer returnAddMoney () {
+        double returnAddMoney () {
             return INT_money;
         }
         public static void main (String[]args){
-
             DepositMoney dialog = new DepositMoney();
             SetPosition.setFrameCenter(dialog);
             dialog.pack();
